requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.27.0
asynctest==0.13.0
awebus==1.0a6
bcrypt==3.2.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.5
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy==1.4.0
mypy-extensions==1.0.0
orjson==3.9.2
packaging==23.1
pluggy==1.2.0
pycparser==2.21
pydantic==2.0.3
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.3.0
PyJWT==2.7.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
tomli==2.0.1
types-python-dateutil==2.8.19.13
types-PyYAML==6.0.12.10
typing_extensions==4.6.3
ujson==5.8.0
uvicorn==0.23.0
watchfiles==0.19.0
websockets==11.0.3
sendgrid==6.10.0
black==23.9.1
pycln==2.2.2
qrcode==7.4.2
freezegun==1.2.2
Pillow==10.1.0
boto3==1.28.65

mercadopago==2.2.3
botocore==1.31.65
types-requests==2.32.0.20241016
